Moved config stuff into their own script
