+++ api/scripts/load_test_data.py
@@ -1,7 +1,7 @@
 import json
 
 from mod.src.database import database, db_models
-from scripts import file_path, meta, session
+from scripts.config import file_path, meta, session
 
 
 def get_class_by_tablename(tablename):
